Add getDatesToDisplay to helpers and remove from Controls
